requirements.txt
pandas==2.2.2
numpy==1.26.4
Faker==25.2.0
PyYAML==6.0.1
scikit-learn==1.5.0
feast==0.40.0
sentence-transformers
pyarrow
tensorflow==2.16.1
tensorflow-recommenders==0.7.3
mlflow==2.13.0
xgboost==2.0.3 # Added for ranking model
lightgbm==4.3.0 # Added for ranking model
joblib==1.4.2 # Often used with scikit-learn models for saving
pymilvus # For Milvus vector database
cassandra-driver==3.29.1 # For ScyllaDB/Cassandra connection 
tf-keras